[ADD] 유저프로필 컴포넌트에서 카테고리 관리
